mnist/mnist-webapp/requirements.txt
flask==0.12.2
grpc==0.3.post19
numpy==1.14.0
Pillow==5.0.0
python-mnist==0.5
tensorflow==1.5.0
tensorflow-serving-api==1.4.0
